Deleted 5 sources under 'tools'
I was checking out the 'tools' section to see if there was anything new and saw quite a few were added.  But when visiting the links I found most did not offer a tool to download or if they did, the copy was limited in nature and had a trial period of under 30 days.  My impression of this list is that it's to be limited to tools people can actually use during their penetration test, not links to companies who's software isn't even publicly available.  If I'm wrong, don't commit.  I just found it really irritating.